renamed video-upload.ts to video.ts
